pyproject.toml
[project]
name = "paddle"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "paddleocr>=3,<4",
    "paddlepaddle==3.3.0",
    "pillow>=12.1.1",
    "pynput>=1.8.1",
]

[[tool.uv.index]]
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
